+++ .eslintrc.js
@@ -40,10 +40,10 @@ module.exports = {
     'no-underscore-dangle': [0],
     'no-plusplus': [0],
     'no-unused-expressions': [0],
-    'no-console': [1, { allow: ['info', 'warn', 'error'] }],
+    'no-console': [0, { allow: ['info', 'warn', 'error'] }],
     'react/prop-types': [2],
     'react/default-props-match-prop-types': [2],
-    'react/forbid-prop-types': [1],
+    'react/forbid-prop-types': [0],
     'react/no-array-index-key': [2],
     'react/no-redundant-should-component-update': [1],
     'react/jsx-props-no-spreading': [0],
